+++ NEWS.md
@@ -1,5 +1,7 @@
 # Changes in mirt 1.8
 
+- `fscores()` nows accepts `method = "plausible"` to draw a single plausible value set
+
 - `plot()` default type is now `score`, and will accept rotation arguments for exploratory models 
   (default rotation is `'none'`)